Delete Answer With Ajax

+++ app/controllers/answers_controller.rb
@@ -14,12 +14,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    if current_user.author?(answer)
-      answer.destroy
-      redirect_to question_path(answer.question), notice: 'Answer successfully deleted.'
-    else
-      redirect_to question_path(answer.question), notice: 'You have no rigths to delete this answer.'
-    end
+    answer.destroy if current_user.author?(answer)
   end
 
   private
